+++ tests/annotations/test_complexity.py
@@ -1,10 +1,14 @@
 """Tests for complexity annotations"""
 
+from typing import Type
 import numpy as np
+from numpy.lib.arraysetops import isin
 import pytest
 
 import pybryt
 import pybryt.complexities as cplx
+
+from pybryt.execution import MemoryFootprintValue
 
 from .utils import assert_object_attrs
 
@@ -14,7 +18,7 @@ def generate_complexity_footprint(name, t_transform, max_exp=8):
     for i, e in enumerate(range(1, max_exp + 1)):
         n = 10 ** e
         t = t_transform(n)
-        values += [pybryt.TimeComplexityResult(name, n, 0, t), i]
+        values.append(MemoryFootprintValue(pybryt.TimeComplexityResult(name, n, 0, t), i, None))
     return pybryt.MemoryFootprint.from_values(*values)
 
 
@@ -84,3 +88,49 @@ def test_time_complexity():
 def test_alias():
     from pybryt.annotations.complexity import complexities as cplx2
     assert cplx.complexity_classes is cplx2.complexity_classes
+
+
+def test_complexity_union():
+    """
+    Tests for complexity unions.
+    """
+    union = cplx.logarithmic | cplx.linear
+    assert isinstance(union, cplx.ComplexityUnion)
+    assert set(union.get_complexities()) == {cplx.logarithmic, cplx.linear}
+
+    union2 = union | cplx.constant
+    assert isinstance(union2, cplx.ComplexityUnion)
+    assert set(union2.get_complexities()) == {cplx.logarithmic, cplx.linear, cplx.constant}
+
+    union3 = cplx.constant | union
+    assert isinstance(union3, cplx.ComplexityUnion)
+    assert set(union3.get_complexities()) == {cplx.logarithmic, cplx.linear, cplx.constant}
+
+    union4 = union2 | union
+    assert isinstance(union4, cplx.ComplexityUnion)
+    assert set(union4.get_complexities()) == {cplx.logarithmic, cplx.linear, cplx.constant}
+
+    union.add_complexity(cplx.linearithmic)
+    assert set(union.get_complexities()) == {cplx.logarithmic, cplx.linear, cplx.linearithmic}
+
+    assert union == cplx.logarithmic | cplx.linear | cplx.linearithmic
+    assert union != cplx.logarithmic | cplx.linear
+
+    # test whether the union works with annotations
+    a = pybryt.TimeComplexity(cplx.constant | cplx.logarithmic, name="foo")
+    footprint = generate_complexity_footprint("foo", lambda v: 1012)
+    res = a.check(footprint)
+    assert res.satisfied
+    assert res.value == cplx.constant
+
+    footprint = generate_complexity_footprint("foo", lambda v: v)
+    res = a.check(footprint)
+    assert not res.satisfied
+    assert res.value == cplx.linear
+
+    # test errors
+    with pytest.raises(TypeError):
+        cplx.constant | 1
+
+    with pytest.raises(TypeError):
+        cplx.ComplexityUnion.from_or(1, cplx.constant)